refactor: utils 도메인 폴더 정리 및 loadable lib 이동
