+++ src/routes/MainRoutes.tsx
@@ -1,12 +1,11 @@
 import { lazy } from "react";
 import { Navigate, type RouteObject } from "react-router-dom";
 
-import { loadable } from "@/utils/loadable";
-
 import WorkspaceListLoading from "@/components/workspace/WorkspaceListLoading";
 
 import RoleGuard from "./RoleGuard";
 
+import { loadable } from "@/lib/loadable";
 import WorkspaceBillingRedirect from "@/pages/workspace/WorkspaceBillingRedirect";
 
 const OverviewDashboard = loadable(
