+++ src/utils/auth/formatPhoneNumber.ts
